Admins, active users. Basics complete! Except equipment loading...

+++ src/io/arkeus/sword/activity/battle/Area.scala
@@ -4,7 +4,7 @@ import scala.util.Random
 import io.arkeus.sword.util.RandomUtils
 
 class Area(val id: Int, val name: String, val monsters: List[Monster]) {
-	val level = (monsters.map(_.level).sum / monsters.length.toDouble).ceil - 1
+	val level = ((monsters.map(_.level).sum / monsters.length.toDouble).ceil - 1).toInt
 	
 	def randomEnemy = monsters(RandomUtils.random(0, monsters.length - 1))
 }
